+++ src/ContainerLayout/index.js
@@ -1,0 +1,2 @@
+import { ContainerLayout } from './ContainerLayout'
+export { ContainerLayout }